requirements.txt
discord.py
psutil
Pillow
aiohttp
cryptography
requests
pyinstaller